+++ framework_lib/src/lib.rs
@@ -38,11 +38,11 @@ pub mod commandline;
 pub mod csme;
 pub mod ec_binary;
 pub mod esrt;
+#[cfg(feature = "uefi")]
+pub mod fw_uefi;
 mod os_specific;
 pub mod power;
 pub mod smbios;
-#[cfg(feature = "uefi")]
-pub mod fw_uefi;
 mod util;
 
 pub mod built_info {
